DRY freeform logic, migrate test
